irc: Change read_socket to return all data it can
irc::IrcCon::read_socket() now returns all the data it can from the
stream.  What needs to be done is to callback to the BufStream to
inform it of how much data we processed.  I also modifed the test to
correctly read from the stream and parse line by line using an
iterator.  This is the model of how main should use it. Might be worth
putting into a function.

TEST=cargo test; all tests pass.