irc: Change read_socket to return a vector.
Previously irc::IrcCon::read_socket() was just returning a string of
what it read from the buffer.  Now, it returns a vector where each
elements represents a line read from the stream.  This was inspired
from the unit test.  It seemed like it was doing too much, now it
simply calls the the function and verifies the input strings.

TEST=cargo test