irc: Verified bytes written and added new test.
I discovered that write_all wasn't actually writing anything out the
wire.  Having changed it to simply write() followed by a flush()
causes bytes to be sent out the wire.  I then used wireshark to
analyze the traffic and discovered that the IRC was indeed sending
back replies, but my client wasn't receiving them.  This led me to
create a 2nd test which does the following:

- Open a file and write two lines to it.
- Call read_socket() which perfoms a BufStream::fill_buf().
- Call update which reports how many bytes are consumed by the buffer.
- Open the file again and write another line.
- Call read_socket()
- Check that you read the last line.

The test currently fails meaning that subsequent calls to fill_buf
aren't actually filling the buf.  I'm not sure if this whole thing is
broken, but I need to find away around this.

TEST=cargo test; tests::read_updated_buffer fails.