main: Fixed off by 1 error in consuming.
I realized that the reason that my test was failing was probably
because the original stream that I wrapped the BufReader around goes
out of scope.  This meant that the BufReader could no longer read from
that stream.  Additionally, I found that the last item in the read
buffer was an empty string.  By changing the functionality such that
we ignore the last empty string.

TEST=cargo run; connected to Rizon and saw the responses printed out
on the console.